pyproject.toml
[tool.poetry]
name = "cert-parser"
version = "0.1.0"
description = "tools for parsing RH entitlement manifests"
authors = ["Stuart Sears <stuart@sjsears.com>"]
license = "GPL-3.0-or-later"
include = ["CHANGELOG.md"]
repository = "https://github.com/lanky/cert-parser"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
rhsm = "^1.19.2"
Jinja2 = "^3.1.2"
pip-tools = "^6.12.2"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pre-commit = "^3.1.0"
pytest = "^7.2.1"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/lanky/cert-parser/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
fixable = ["ALL"]
select = ["E", "F", "UP", "B", "SIM", "I"]
extend-select = ["B", "I"]
extend-fixable = ["I"]
ignore = ["B006", "B904"]
#
[tool.ruff.format]
docstring-code-format = true
preview = true
#
[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
suppress-dummy-args = true
